Refactored front end compoennts
