--- front/src/components/spellbookContainer/SpellbookContainer.module.css
@@ -1,5 +1,0 @@
-.spellbookContainer {
-    display: flex;
-    flex-direction: column;
-    gap: 2rem;
-}
