+++ front/src/components/spellbookTables/SpellbookTables.module.css
@@ -1,0 +1,7 @@
+.spellbookTables {
+    margin-bottom: 4rem;
+    display: flex;
+    flex-direction: column;
+    align-items: start;
+    gap: 1rem;
+}
